+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 * Updated documentation
 * Fixed a bug that prevented thumbnail to be shown
 * Fixed blurry text on some screens with DPI scaling
+* added info to format selector